- Swift's Swizzling was refactored - Objective-C's Swizzling was improved
