+++ Tests/NSObject+Swizzling.h
@@ -17,11 +17,6 @@
 
 @interface NSObject (Swizzling)
 
-+ (void)exchangeClassSelector:(SEL)originalSelector
-                 withSelector:(SEL)swizzledSelector;
-
-+ (void)exchangeInstanceCls1:(Class)cls1 Sel1:(SEL)sel1 Cls2:(Class)cls2 Sel2:(SEL)sel2;
-
 /**
  *  Replaces the selector's associated method implementation with the
  *  given implementation (or adds it, if there was no existing one).
